implement "develop" command
The develop command is used like this:

    hit develop [target]

The [target] is optional, but allows you to quickly rename your default
profile to something else for testing/work.

By default, the development profile is created using absolute symbolic
links back to the original builds in your HashDist build store.  There
is weak support for relative symbolic links and files copies, but these
are not well-tested at this point.

This commit also reverts cefa0b7c4a1f1e7ac439c91d57268bdb70382f4c, which
was creating blanket write-protection on profiles.  Individual packages
should now be write-protected, so users should not be able to corrupt
their installs by attempting to modify a development profile.

Multiline shebangs and launcher appear to be somewhat confused, so
please consider this commit to be a work-in-progress.