Add default profiles and use flags
Enable the hashdist config file to specify a default profile that
is inherited by all profiles.  By default, a few common parameters such
as platform are set.

Also, added a new "default_use" section to profiles.  This specifies a
preference for how to resolve common packages.  The most common example
is the blas library, which uses a host blas installation on Linux or
Cygwin, but requires the Accelerate framework on OS X.

Finally, the default platform name for Darwin is now darwin, and the
default platform name for Cygwin is now cygwin.  The default platform
name of Linux has not changed.