hashdist.cli.build_tools_cli: PEP 257 docstring for BuildPostprocess
Use a phrase ending with a period [1], followed by a blank line and
additional details [2].  This avoids 'hit --help' output like:

    build-postprocess   Walks through directories to perform the actions given
                        by flags
    build-unpack-sources
                        Extracts ...

which cut off in mid-sentence.  Now it cuts off in the same place, but
it's clear on the docstring side that it's the end of the sentence ;).

I also switched to imperitive 'Walk' instead of the implicit
description '[This command] Walks' [1].

[1]: http://legacy.python.org/dev/peps/pep-0257/#one-line-docstrings
[2]: http://legacy.python.org/dev/peps/pep-0257/#multi-line-docstrings